requirements.txt
asgiref==3.8.1
certifi==2024.8.30
charset-normalizer==3.4.0
cloudinary==1.41.0
dj-database-url==2.3.0
Django==5.1.2
django-autoslug==1.9.9
django-cloudinary-storage==0.3.0
django-crispy-forms==2.3
django-filter==24.3
django-pagination==1.0.7
django-taggit==6.1.0
django-tinymce==4.1.0
django-widget-tweaks==1.5.0
gunicorn==23.0.0
idna==3.10
packaging==24.2
pillow==11.0.0
psycopg2-binary==2.9.10
python-dotenv==1.0.1
requests==2.32.3
six==1.17.0
sqlparse==0.5.2
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
